requirements.txt
matplotlib==3.2.2
numpy==1.18.1
Pillow==7.0.0
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
tqdm==4.62.3
